Recoded `Workstep` for better parameter handling
This commit adds the `WorkstepUsage` class and its dependencies. This allows better handling of the paramters for each `WorkstepType`.

Breaking Changes
The class `Workstep` will be changed in near future, at the moment, both ways are possible to avoid breaking existings app. It is recommended to switch to `WorkstepUsage` instead of using `WorkstepDuration`. Later will be removed in an upcoming release.

Fixed #80